remove generic rle
