+++ vortex-btrblocks/src/schemes/mod.rs
@@ -12,4 +12,56 @@ pub mod decimal;
 pub mod temporal;
 
 pub(crate) mod patches;
-pub(crate) mod rle;
+
+use vortex_compressor::builtins::FloatDictScheme;
+use vortex_compressor::builtins::IntDictScheme;
+use vortex_compressor::builtins::StringDictScheme;
+use vortex_compressor::scheme::AncestorExclusion;
+use vortex_compressor::scheme::ChildSelection;
+use vortex_compressor::scheme::DescendantExclusion;
+use vortex_compressor::scheme::SchemeExt;
+
+use crate::schemes::integer::SparseScheme;
+
+/// Shared descendant exclusion rules for RLE schemes.
+///
+/// RLE indices (child 1) and offsets (child 2) are monotonically increasing positions with all
+/// unique values. Dict and Sparse are pointless on such data. Self-exclusion already prevents
+/// RLE on RLE children.
+fn rle_descendant_exclusions() -> Vec<DescendantExclusion> {
+    vec![
+        DescendantExclusion {
+            excluded: IntDictScheme.id(),
+            children: ChildSelection::Many(&[1, 2]),
+        },
+        // TODO(connor): This is wrong for some reason?
+        // DescendantExclusion {
+        //     excluded: RunEndScheme.id(),
+        //     children: ChildSelection::Many(&[1, 2]),
+        // },
+        DescendantExclusion {
+            excluded: SparseScheme.id(),
+            children: ChildSelection::Many(&[1, 2]),
+        },
+    ]
+}
+
+/// Shared ancestor exclusion rules for RLE schemes.
+///
+/// Dict values (child 0) are all unique by definition, so RLE is pointless on them.
+fn rle_ancestor_exclusions() -> Vec<AncestorExclusion> {
+    vec![
+        AncestorExclusion {
+            ancestor: IntDictScheme.id(),
+            children: ChildSelection::One(0),
+        },
+        AncestorExclusion {
+            ancestor: FloatDictScheme.id(),
+            children: ChildSelection::One(0),
+        },
+        AncestorExclusion {
+            ancestor: StringDictScheme.id(),
+            children: ChildSelection::One(0),
+        },
+    ]
+}
